+++ ploomes-teste.persistence/Contracts/ITipoLugarRepository.cs
@@ -1,8 +1,10 @@
+using ploomes_teste.domain;
+
 namespace ploomes_teste.persistence.Contracts
 {
     public interface ITipoLugarRepository : IGeneralRepository
     {
-        Task<TipoLugar> GetTipoLugarById(int id);
+        Task<TipoLugar> GetTipoLugarById(short id);
         Task<TipoLugar[]> GetTiposLugar();
     }
 }
